requirements.txt
asgiref==3.6.0
Django==3.2.9
django-ckeditor==6.5.1
django-js-asset==2.0.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
